feat: 完成 Fortran 95 到 Fortran 2023 全面现代化
主要变更:
- 🔄 类型系统现代化: REAL*8 → real(real64), DOUBLE PRECISION → real(real64)
- 📦 模块系统现代化: 直接使用 iso_fortran_env 替代 vast_kind_param
- 🗂️ 按依赖顺序现代化: libmod → librang90 → libmcp90 → libdvd90 → lib9290 → ris4
- 🧹 清理历史包袱: 删除过时的兼容性模块和工具

技术修复:
- 修复 dvdson.f90 大型多子程序文件现代化
- 解决 iniest2.f90 数组秩不匹配问题 (EIGVAL → EIGVAL(1))
- 处理重复 iso_fortran_env 导入问题
- 配置 FlexiBLAS 支持解决 BLAS 库检测

构建结果:
- ✅ 成功编译所有库: libmod.a, libdvd90.a, lib9290.a, libmcp90.a, librang90.a
- ✅ 成功构建 ris4 可执行文件
- 📈 处理 600+ 个 .f90 文件的现代化

删除文件:
- 移除过时的 vast_kind_param_M.f90 模块
- 清理未使用的 mpi90 和 tool 目录
- 删除 relativistic_isotope_shift.f90 (由 isotope_shift.f90 替代)

添加文件:
- FORTRAN_MODERNIZATION_GUIDE.md: 完整的现代化指南和技术文档
- isotope_shift.f90: 现代化的相对论同位素位移计算程序

🎯 目标达成: 消除所有 "历史包袱", 全面采用现代 Fortran 2023 标准

+++ src/appl/ris4/smsread.f90
@@ -26,7 +26,7 @@
 !-----------------------------------------------
 !   D u m m y   A r g u m e n t s
 !-----------------------------------------------
-      REAL(DOUBLE), DIMENSION(NNNW,NNNW), INTENT(IN) :: VINT, VINT2
+      real(real64), DIMENSION(NNNW,NNNW), INTENT(IN) :: VINT, VINT2
 !-----------------------------------------------
 !   L o c a l   P a r a m e t e r s
 !-----------------------------------------------
@@ -34,7 +34,7 @@
 !-----------------------------------------------
 !   L o c a l   V a r i a b l e s
 !-----------------------------------------------
-      REAL(DOUBLE) :: CONTRI, CONTRIK1, COEFFSMS
+      real(real64) :: CONTRI, CONTRIK1, COEFFSMS
       INTEGER :: IC, IR, I, J, IIA, IIB, IIC, IID, LOC, LAB, IOS
 !-----------------------------------------------
 !
